+++ src/java/com/twitter/common/args/ArgumentInfo.java
@@ -66,6 +66,7 @@ public abstract class ArgumentInfo<T> {
   private final String canonicalName;
   private final String name;
   private final String help;
+  private final boolean argFile;
   private final Arg<T> arg;
   private final TypeToken<T> type;
   private final List<Annotation> verifierAnnotations;
@@ -77,6 +78,7 @@ public abstract class ArgumentInfo<T> {
    * @param canonicalName A fully qualified name for the argument.
    * @param name The simple name for the argument.
    * @param help Help string.
+   * @param argFile If argument file is allowed.
    * @param arg Argument object.
    * @param type Concrete argument type.
    * @param verifierAnnotations {@link com.twitter.common.args.Verifier} annotations for this
@@ -87,6 +89,7 @@ public abstract class ArgumentInfo<T> {
       String canonicalName,
       String name,
       String help,
+      boolean argFile,
       Arg<T> arg,
       TypeToken<T> type,
       List<Annotation> verifierAnnotations,
@@ -95,6 +98,7 @@ public abstract class ArgumentInfo<T> {
     this.canonicalName = MorePreconditions.checkNotBlank(canonicalName);
     this.name = MorePreconditions.checkNotBlank(name);
     this.help = MorePreconditions.checkNotBlank(help);
+    this.argFile = argFile;
     this.arg = Preconditions.checkNotNull(arg);
     this.type = Preconditions.checkNotNull(type);
     this.verifierAnnotations = ImmutableList.copyOf(verifierAnnotations);
@@ -126,6 +130,13 @@ public abstract class ArgumentInfo<T> {
    */
   public String getHelp() {
     return help;
+  }
+
+  /**
+   * Returns whether an argument file is allowed for this argument.
+   */
+  public boolean argFile() {
+    return argFile;
   }
 
   /**